backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.29.0
apscheduler==3.10.4
aiogram==3.13.0
python-dotenv==1.0.1
pytz==2024.2
aiofiles==24.1.0
